chore: move github code to github module
