+++ cargo-dist/src/tasks.rs
@@ -49,7 +49,6 @@
 //! steps to give them the freedom to do whatever they need to do.
 
 use std::collections::BTreeMap;
-use std::fs::File;
 
 use axoasset::AxoClient;
 use axoprocess::Cmd;
@@ -57,7 +56,7 @@ use axoproject::platforms::{
     TARGET_ARM64_LINUX_GNU, TARGET_ARM64_MAC, TARGET_X64_LINUX_GNU, TARGET_X64_MAC,
 };
 use axoproject::{PackageId, PackageIdx, WorkspaceGraph};
-use camino::{Utf8Path, Utf8PathBuf};
+use camino::Utf8PathBuf;
 use cargo_dist_schema::{ArtifactId, DistManifest, GithubJobStep, SystemId, SystemInfo};
 use semver::Version;
 use serde::Serialize;
@@ -1124,7 +1123,10 @@ impl<'pkg_graph> DistGraphBuilder<'pkg_graph> {
         )?;
         let github_build_setup = github_build_setup
             .as_ref()
-            .map(|local| GithubJobStepsBuilder::new(&workspace_dir, local)?.validate())
+            .map(|local| {
+                crate::backend::ci::github::GithubJobStepsBuilder::new(&workspace_dir, local)?
+                    .validate()
+            })
             .transpose()?
             .unwrap_or_default();
         Ok(Self {
@@ -3125,301 +3127,5 @@ fn submodule_head(submodule_path: &Utf8PathBuf) -> DistResult<Option<String>> {
         Ok(None)
     } else {
         Ok(Some(commit.to_owned()))
-    }
-}
-
-/// Builder for looking up and reporting errors in the steps provided by the
-/// `github-build-setup` configuration
-pub struct GithubJobStepsBuilder {
-    steps: Vec<GithubJobStep>,
-    path: Utf8PathBuf,
-}
-
-impl GithubJobStepsBuilder {
-    #[cfg(test)]
-    /// Test only ctor for skipping the fs lookup
-    pub fn from_values(
-        steps: impl IntoIterator<Item = GithubJobStep>,
-        path: impl Into<Utf8PathBuf>,
-    ) -> Self {
-        Self {
-            steps: Vec::from_iter(steps.into_iter()),
-            path: path.into(),
-        }
-    }
-
-    /// Create a new validator
-    pub fn new(
-        base_path: impl AsRef<Utf8Path>,
-        cfg_value: impl AsRef<Utf8Path>,
-    ) -> Result<Self, DistError> {
-        let path = base_path.as_ref().join(cfg_value.as_ref());
-        let mut f = File::options().read(true).open(&path).map_err(|e| {
-            eprintln!("Error opening {path}: {e}");
-            crate::DistError::GithubBuildSetupNotFound {
-                expected_path: path.clone(),
-            }
-        })?;
-        let steps = serde_yml::from_reader(&mut f)?;
-        Ok(Self {
-            steps,
-            path: path.to_path_buf(),
-        })
-    }
-
-    /// Validate the whole list of build setup steps
-    pub fn validate(self) -> Result<Vec<GithubJobStep>, DistError> {
-        for (i, step) in self.steps.iter().enumerate() {
-            if let Some(message) = Self::validate_step(i, step) {
-                return Err(DistError::GithubBuildSetupNotValid {
-                    file_path: self.path.to_path_buf(),
-                    message,
-                });
-            }
-        }
-        Ok(self.steps)
-    }
-
-    /// validate a single step in the list of steps, returns `Some` if an error is detected
-    fn validate_step(idx: usize, step: &GithubJobStep) -> Option<String> {
-        //github-build-step {step_name} is invalid, cannot have both `uses` and `{conflict_step_name}` defined
-        let key_mismatch = |lhs: &str, rhs: &str| {
-            let step_name = Self::get_name_id_or_idx(idx, step);
-            format!("github-build-step {step_name} is invalid, cannot have both `{lhs}` and `{rhs}` defined")
-        };
-        let invalid_object = |prop: &str, msg: &str| {
-            let step_name = Self::get_name_id_or_idx(idx, step);
-            format!("github-build-step {step_name} has an invalid `{prop}` entry: {msg}")
-        };
-        if let Some(key) = Self::validate_step_uses_keys(step) {
-            return Some(key_mismatch("uses", key));
-        }
-        if let Some(key) = Self::validate_step_run_keys(step) {
-            return Some(key_mismatch("run", key));
-        }
-        if let Some(message) = Self::validate_with_shape(step) {
-            return Some(invalid_object("with", &message));
-        }
-        None
-    }
-
-    /// Validate there are no conflicting keys in this workflow that defines the `uses` keyword
-    ///
-    /// returns `Some` if an error is detected
-    fn validate_step_uses_keys(step: &GithubJobStep) -> Option<&'static str> {
-        // if uses is None, return early
-        step.uses.as_ref()?;
-        if step.run.is_some() {
-            return Some("run");
-        }
-        if step.shell.is_some() {
-            return Some("shell");
-        }
-        if step.working_directory.is_some() {
-            return Some("working-directory");
-        }
-        None
-    }
-
-    /// Validate there are no conflicting keys in this workflow that defines the `run` keyword
-    /// this is called _after_ `Self::validate_step_uses_keys`
-    ///
-    /// returns `Some` if an error is detected
-    fn validate_step_run_keys(step: &GithubJobStep) -> Option<&'static str> {
-        // if run is None, return early
-        step.run.as_ref()?;
-        if !step.with.is_empty() {
-            return Some("with");
-        }
-        None
-    }
-
-    /// Validate the with mapping only contains key/value pairs with the values being either
-    /// strings, booleans, or numbers
-    ///
-    /// returns `Some` if an error is detected
-    fn validate_with_shape(step: &GithubJobStep) -> Option<String> {
-        for (k, v) in &step.with {
-            let invalid_type = match v {
-                serde_json::Value::Null => "null",
-                serde_json::Value::Array(_) => "array",
-                serde_json::Value::Object(_) => "object",
-                _ => continue,
-            };
-            return Some(format!("key `{k}` has the type of `{invalid_type}` only `string`, `number` or `boolean` are supported"));
-        }
-        None
-    }
-
-    fn get_name_id_or_idx(idx: usize, step: &GithubJobStep) -> String {
-        step.name
-            .clone()
-            .or_else(|| step.id.clone())
-            .unwrap_or_else(|| idx.to_string())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use serde_json::Value;
-
-    use super::*;
-
-    #[test]
-    fn validator_works() {
-        let steps = [GithubJobStep {
-            uses: Some("".to_string()),
-            with: BTreeMap::from_iter([
-                ("key".to_string(), Value::from("value")),
-                ("key2".to_string(), Value::from(2)),
-                ("key2".to_string(), Value::from(false)),
-            ]),
-            timeout_minutes: Some("8".into()),
-            continue_on_error: Some("true".into()),
-            ..Default::default()
-        }];
-        let path = Utf8PathBuf::from(std::thread::current().name().unwrap_or(""));
-        GithubJobStepsBuilder::from_values(steps, &path)
-            .validate()
-            .expect("validation to pass");
-    }
-
-    #[test]
-    #[should_panic = "cannot have both `uses` and `run` defined"]
-    fn validator_catches_run_and_uses() {
-        let steps = [GithubJobStep {
-            uses: Some("".to_string()),
-            run: Some("".to_string()),
-            ..Default::default()
-        }];
-        let path = Utf8PathBuf::from(std::thread::current().name().unwrap_or(""));
-        GithubJobStepsBuilder::from_values(steps, &path)
-            .validate()
-            .unwrap();
-    }
-
-    #[test]
-    #[should_panic = "cannot have both `uses` and `shell` defined"]
-    fn validator_catches_run_and_shell() {
-        let steps = [GithubJobStep {
-            uses: Some("".to_string()),
-            shell: Some("".to_string()),
-            ..Default::default()
-        }];
-        let path = Utf8PathBuf::from(std::thread::current().name().unwrap_or(""));
-        GithubJobStepsBuilder::from_values(steps, &path)
-            .validate()
-            .unwrap();
-    }
-
-    #[test]
-    #[should_panic = "cannot have both `uses` and `working-directory` defined"]
-    fn validator_catches_run_and_cwd() {
-        let steps = [GithubJobStep {
-            uses: Some("".to_string()),
-            working_directory: Some("".to_string()),
-            ..Default::default()
-        }];
-        let path = Utf8PathBuf::from(std::thread::current().name().unwrap_or(""));
-        GithubJobStepsBuilder::from_values(steps, &path)
-            .validate()
-            .unwrap();
-    }
-
-    #[test]
-    #[should_panic = "cannot have both `run` and `with` defined"]
-    fn validator_catches_run_and_with() {
-        let steps = [GithubJobStep {
-            run: Some("".to_string()),
-            with: BTreeMap::from_iter([("key".to_string(), Value::from("value"))]),
-            ..Default::default()
-        }];
-        let path = Utf8PathBuf::from(std::thread::current().name().unwrap_or(""));
-        GithubJobStepsBuilder::from_values(steps, &path)
-            .validate()
-            .unwrap();
-    }
-
-    #[test]
-    #[should_panic = "has an invalid `with` entry: key `key` has the type of `object` only `string`, `number` or `boolean` are supported"]
-    fn validator_catches_invalid_with() {
-        let steps = [GithubJobStep {
-            uses: Some("".to_string()),
-            with: BTreeMap::from_iter([(
-                "key".to_string(),
-                serde_json::json!({
-                    "obj-key": "obj-value"
-                }),
-            )]),
-            ..Default::default()
-        }];
-        let path = Utf8PathBuf::from(std::thread::current().name().unwrap_or(""));
-        GithubJobStepsBuilder::from_values(steps, &path)
-            .validate()
-            .unwrap();
-    }
-
-    #[test]
-    #[should_panic = "step-name"]
-    fn validator_errors_with_name() {
-        let steps = [GithubJobStep {
-            name: Some("step-name".to_string()),
-            uses: Some(String::new()),
-            run: Some(String::new()),
-            ..Default::default()
-        }];
-        let path = Utf8PathBuf::from(std::thread::current().name().unwrap_or(""));
-        GithubJobStepsBuilder::from_values(steps, &path)
-            .validate()
-            .unwrap();
-    }
-
-    #[test]
-    #[should_panic = "step-name"]
-    fn validator_errors_with_name_over_id() {
-        let steps = [GithubJobStep {
-            name: Some("step-name".to_string()),
-            id: Some("step-id".to_string()),
-            uses: Some(String::new()),
-            run: Some(String::new()),
-            ..Default::default()
-        }];
-        let path = Utf8PathBuf::from(std::thread::current().name().unwrap_or(""));
-        GithubJobStepsBuilder::from_values(steps, &path)
-            .validate()
-            .unwrap();
-    }
-
-    #[test]
-    #[should_panic = "step-id"]
-    fn validator_errors_with_id() {
-        let steps = [GithubJobStep {
-            id: Some("step-id".to_string()),
-            uses: Some(String::new()),
-            run: Some(String::new()),
-            ..Default::default()
-        }];
-        let path = Utf8PathBuf::from(std::thread::current().name().unwrap_or(""));
-        GithubJobStepsBuilder::from_values(steps, &path)
-            .validate()
-            .unwrap();
-    }
-
-    #[test]
-    fn build_setup_can_read() {
-        let tmp = temp_dir::TempDir::new().unwrap();
-        let base = Utf8PathBuf::from_path_buf(tmp.path().to_owned())
-            .expect("temp_dir made non-utf8 path!?");
-        let cfg = "build-setup.yml".to_string();
-        std::fs::write(
-            &base.join(&cfg),
-            r#"
-- uses: some-action-user/some-action
-  continue-on-error: ${{ some.expression }}
-  timeout-minutes: ${{ matrix.timeout }}
-"#,
-        )
-        .unwrap();
-        GithubJobStepsBuilder::new(&base, &cfg).unwrap();
     }
 }
